% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_params}
\alias{read_params}
\title{Import AHLE parameter file}
\usage{
read_params(file_path, file_type = "yaml")
}
\arguments{
\item{file_path}{Relative or absolute path to file}

\item{file_type}{File type (One of: YAML (default), JSON, or CSV)}
}
\description{
Read in AHLE parameter file in YAML, JSON, or CSV format.
}
