% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpert.R
\name{rpert}
\alias{rpert}
\title{Generate random variates from PERT distribution for AHLE modelling}
\usage{
rpert(n, x.min, x.max, x.mode, lambda = 4)
}
\arguments{
\item{n}{Number of variates to generate}

\item{x.min}{Minimum x value to sample}

\item{x.max}{Maximum x value to sample}

\item{x.mode}{Most frequent x value to sample}

\item{lambda}{Scaling factor (set to 4 by default)}
}
\value{
A sample of random variates
}
\description{
Returns random variates sampled from a modified PERT distribution given by the following parameterization

\eqn{\mu = \frac{x_{\mathrm{min}} + x_{\mathrm{max}} + \lambda x_{\mathrm{mode}}}{\lambda + 2}}

If \eqn{\mu = x_{\mathrm{mode}}}, then \eqn{v = \frac{\lambda}{2} + 1}, else \eqn{v = \frac{(\mu - x_\mathrm{min})(2x_\mathrm{mode} - x_\mathrm{min} - x_\mathrm{max})}{(x_\mathrm{mode} - \mu)(x_\mathrm{max} - x_\mathrm{min})}}

Define \eqn{w = \frac{v(x_\mathrm{max} - \mu)}{mu - x_\mathrm{min}}}, then generate variates from a \eqn{Beta(v, w)x_\mathrm{range} + x_\mathrm{min}} distribution, where \eqn{x_\mathrm{range} = x_\mathrm{max} - x_\mathrm{min}}
}
