% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params.R
\name{read_params}
\alias{read_params}
\title{Import DPM/AHLE parameter file}
\usage{
read_params(file_path, file_type = "yaml")
}
\arguments{
\item{file_path}{Relative or absolute path to file}

\item{file_type}{File type (Currently only YAML is supported)}
}
\description{
Read in DPM/AHLE parameter file in YAML format
}
