% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpert.R
\name{rpert}
\alias{rpert}
\title{Generate random variates from PERT distribution}
\usage{
rpert(n, x.min, x.max, x.mode, lambda = 4)
}
\arguments{
\item{n}{Number of variates to generate}

\item{x.min}{Minimum x value to sample}

\item{x.max}{Maximum x value to sample}

\item{x.mode}{Most frequent x value to sample}

\item{lambda}{Scaling factor}
}
\value{
A sample of random variates
}
\description{
Returns random variates sampled from a modified PERT distribution given by 
the following paramterization
}
\examples{
# rpert(10000, 1, 10, 5)

}
