% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the Animal Health Loss Envelope (AHLE) compartmental agent-based simulation model}
\usage{
run_model()
}
\description{
Runs simulations using parameters imported via \code{read_params()}
}
