% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up and run DPM/AHLE scenarios}
\usage{
setup(file_path, seed_value = NULL)
}
\arguments{
\item{file_path}{A path to a directory containing YAML files to be analyzed}

\item{seed_value}{An integer value of any length for reproducibility}
}
\description{
Read in a DPM/AHLE parameter file in YAML format and run compartmental model.
}
